# spec/openapi/ai_agent_platform_openapi_v1.yaml
openapi: 3.1.0
info:
  title: MSSQL Analysis Agent Platform API
  version: 0.1.0
  summary: OpenAPI contract skeleton for the MSSQL analysis, documentation, and Java/MyBatis draft-generation platform.
  description: |
    Starter API contract for the central Agent platform that analyzes MSSQL Stored Procedures
    and related database metadata, produces documentation, and creates draft Java/MyBatis
    conversion artifacts.

    This contract is intentionally skeleton-level. It defines the stable API surface and
    response shapes expected by the first implementation wave while preserving project
    policies:

    - metadata access is read-only and routed through the MSSQL Metadata MCP boundary;
    - generated documentation, Java/MyBatis code, DTO/VO/Model, and DDL are draft artifacts;
    - default P25 workflows stop at validation-complete draft output without user review UI;
    - deferred publish-like actions still require validation evidence and retained approval capability;
    - live row-data queries, automatic DDL execution, and unreviewed code application are out of scope.
servers:
  - url: http://localhost:{port}
    description: Local API/BFF server
    variables:
      port:
        default: "8000"
        description: APP_PORT resolved by Makefile or scripts/resolve_dev_ports.sh.
...
paths:
...
  /api/v1/metadata/tools/{toolName}/invoke:
    post:
      tags: [metadata]
      summary: Invoke a public read-only MSSQL Metadata MCP dependency evidence tool.
      operationId: invokeMetadataTool
      description: |
        Invokes only the public P28 dependency-evidence allowlist through the MSSQL Metadata
        MCP registry boundary: get_dependency_closure and resolve_dependency_reference.
        The request accepts a structured arguments object only. The response is a sanitized
        metadata envelope with snapshot and evidence refs; it never returns row data, executes
        stored procedures, exposes SQL definition text, stores raw metadata, performs DDL/DML,
        or falls back from PPM to PLF.
      parameters:
        - name: toolName
          in: path
          required: true
          schema:
            type: string
            enum:
              - get_dependency_closure
              - resolve_dependency_reference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetadataToolInvokeRequest"
      responses:
        "200":
          description: Sanitized metadata tool invocation envelope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataToolInvokeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "424":
          $ref: "#/components/responses/DependencyBlocked"
        "503":
          $ref: "#/components/responses/DependencyBlocked"
  /api/v1/metadata/search:
    get:
      tags: [metadata]
      summary: Search read-only MSSQL metadata objects.
      operationId: searchMetadataObjects
      description: |
        Searches metadata object identities through the MSSQL Metadata MCP read-only boundary.
        The response is limited to object identity, profile/database context, snapshot/evidence
        references, caveats, review-required markers, and blocker codes. It never returns row
        data, executes stored procedures, exposes SQL definition text, or performs DDL/DML.
        PPM search must not fall back to PLF when PPM is unavailable.
      parameters:
        - name: dbProfileId
          in: query
          required: true
          schema:
            type: string
          examples:
            master:
              value: master
            ppm:
              value: ppm
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          example: order
        - name: objectTypes
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: "#/components/schemas/MetadataSearchObjectType"
            default: [PROCEDURE, TABLE, VIEW, FUNCTION]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Metadata object search results and blocker/caveat context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataSearchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "424":
          $ref: "#/components/responses/DependencyBlocked"
        "503":
          $ref: "#/components/responses/DependencyBlocked"
  /api/v1/metadata/analyze:
    post:
      tags: [metadata]
      summary: Analyze read-only MSSQL metadata with bounded AI-MCP orchestration.
      operationId: analyzeMetadata
      description: |
        Runs response-only metadata analysis. The API first builds deterministic metadata
        identity context from a query or target, then allows the LLM to plan bounded
        active/read-only MCP metadata tool calls. Tool execution stays internal to the API
        registry boundary; the public metadata tool invocation allowlist is not expanded.
        Responses contain sanitized evidence digests, deterministic fact ids, LLM insights,
        review markers, and caveats. They never include row data, raw SQL definition text,
        procedure execution output, DDL/DML controls, secrets, or raw provider traces.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetadataAnalysisRequest"
      responses:
        "200":
          description: Response-only metadata analysis with sanitized AI-MCP evidence.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetadataAnalysisResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "424":
          $ref: "#/components/responses/DependencyBlocked"
        "503":
          $ref: "#/components/responses/DependencyBlocked"
...
components:
...
  responses:
    BadRequest:
      description: Invalid request payload or unsupported workflow option.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The requested read-only metadata dependency is blocked by policy or access.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    DependencyBlocked:
      description: Required metadata dependency is unavailable or blocked by policy.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    EvidenceRef:
      type: object
      required: [type, objectRef, locator]
      properties:
        type:
          type: string
          enum: [MSSQL_METADATA, STATIC_ANALYSIS, LLM_INFERENCE, POLICY, TEMPLATE, USER_INPUT]
        objectRef:
          type: string
          examples: [plf.dbo.usp_OrderRequest_Select]
        locator:
          type: string
        snapshotId:
          type: string
...
    ModelInvocationSummary:
      type: object
      required:
        - provider
        - model
        - modelProfileId
        - promptVersion
        - outputSchemaVersion
        - inputHash
        - promptHash
        - outputHash
        - status
      properties:
        provider:
          type: string
        model:
          type: string
        modelProfileId:
          type: string
        modelRegistryRef:
          type: string
        reasoningEffort:
          type: string
        promptVersion:
          type: string
        outputSchemaVersion:
          type: string
        inputHash:
          type: string
        promptHash:
          type: string
        outputHash:
          type: string
        status:
          type: string
          enum: [SUCCEEDED, FAILED, SKIPPED]
        tokenUsage:
          type: object
          additionalProperties:
            type: integer
        latencyMs:
          type: integer
        componentInvocations:
          type: array
          description: Sanitized per-stage model invocation summaries for staged LLM analysis.
          items:
            type: object
            additionalProperties: true
...
    MetadataToolInvokeRequest:
      type: object
      required: [arguments]
      additionalProperties: false
      properties:
        arguments:
          type: object
          additionalProperties: true
          description: Structured MCP tool arguments only; free-form SQL/write command keys are rejected.
    MetadataToolInvokeResponse:
      type: object
      required: [ok, toolName, dbProfileId, snapshotId, collectedAt, evidenceRefs, data]
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        toolName:
          type: string
          enum:
            - get_dependency_closure
            - resolve_dependency_reference
        dbProfileId:
          type: string
        snapshotId:
          type: string
        collectedAt:
          type: string
          format: date-time
        evidenceRefs:
          type: array
          items:
            type: object
            additionalProperties: true
        data:
          type: object
          additionalProperties: true
    MetadataSearchObjectType:
      type: string
      enum: [PROCEDURE, TABLE, VIEW, FUNCTION]
    MetadataObjectIdentity:
      type: object
      required: [schema, name, type]
      properties:
        schema:
          type: string
        name:
          type: string
        type:
          $ref: "#/components/schemas/MetadataSearchObjectType"
    MetadataSearchBlocker:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples:
            - PPM_MANIFEST_TEMPLATE_ONLY
            - METADATA_SEARCH_MCP_TOOL_MISSING
            - PPM_DB_NOT_FOUND
            - PPM_DB_ACCESS_DENIED
            - LIVE_METADATA_UNAVAILABLE
            - DEPENDENCY_METADATA_INCOMPLETE
            - P21_LIVE_PPM_REQUIRED
        message:
          type: string
    MetadataSearchResult:
      type: object
      required:
        - objectIdentity
        - sourceProfile
        - sourceDatabase
        - evidenceRefs
        - caveats
        - reviewRequired
        - blockers
      properties:
        objectIdentity:
          $ref: "#/components/schemas/MetadataObjectIdentity"
        sourceProfile:
          type: string
        sourceDatabase:
          type: string
        snapshotId:
          type: string
        evidenceRefs:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceRef"
        caveats:
          type: array
          items:
            type: string
        reviewRequired:
          type: boolean
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchBlocker"
    MetadataSearchResponse:
      type: object
      required:
        - dbProfileId
        - query
        - objectTypes
        - limit
        - sourceProfile
        - sourceDatabase
        - results
        - caveats
        - reviewRequired
        - blockers
      properties:
        dbProfileId:
          type: string
        query:
          type: string
        objectTypes:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchObjectType"
        limit:
          type: integer
          minimum: 1
          maximum: 100
        sourceProfile:
          type: string
        sourceDatabase:
          type: string
        snapshotId:
          type: string
        collectedAt:
          type: string
          format: date-time
        results:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchResult"
        caveats:
          type: array
          items:
            type: string
        reviewRequired:
          type: boolean
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchBlocker"
    MetadataAnalysisOptions:
      type: object
      properties:
        useLlmAnalysis:
          type: boolean
          default: true
        useAiToolOrchestration:
          type: boolean
          default: true
        llmProfileId:
          type: string
          default: openai_sp_semantic_analysis
          enum:
            - openai_sp_semantic_analysis
            - openai_fast_test
        maxTargets:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
    MetadataAnalysisRequest:
      type: object
      required: [dbProfileId]
      description: Provide exactly one of query or target.
      properties:
        dbProfileId:
          type: string
        query:
          type: string
          minLength: 1
        target:
          $ref: "#/components/schemas/MetadataObjectIdentity"
        objectTypes:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchObjectType"
          default: [PROCEDURE, TABLE, VIEW, FUNCTION]
        options:
          $ref: "#/components/schemas/MetadataAnalysisOptions"
    MetadataAnalysisInsight:
      type: object
      required: [code, objectRef, summary, status, evidenceRefs]
      properties:
        code:
          type: string
        objectRef:
          type: string
        summary:
          type: string
        status:
          type: string
          enum: [INFERRED_DESCRIPTION, REVIEW_REQUIRED]
        evidenceRefs:
          type: array
          items:
            type: string
    MetadataObjectProfile:
      type: object
      required:
        - objectRef
        - objectType
        - columnCount
        - primaryKeyCount
        - foreignKeyCount
        - indexCount
        - constraintCount
        - descriptionCoverage
        - reviewRequired
        - evidenceRefs
        - sourceFactIds
      properties:
        objectRef:
          type: string
        objectType:
          type: string
        columnCount:
          type: integer
          minimum: 0
        primaryKeyCount:
          type: integer
          minimum: 0
        foreignKeyCount:
          type: integer
          minimum: 0
        indexCount:
          type: integer
          minimum: 0
        constraintCount:
          type: integer
          minimum: 0
        descriptionCoverage:
          type: number
          minimum: 0
          maximum: 1
        reviewRequired:
          type: boolean
        evidenceRefs:
          type: array
          items:
            type: string
        sourceFactIds:
          type: array
          items:
            type: string
    MetadataInsightGroup:
      type: object
      required: [category, insights]
      properties:
        category:
          type: string
          enum:
            - COLUMN_RISK
            - RELATIONSHIP
            - INDEX
            - CONSTRAINT
            - DOCUMENTATION_GAP
            - DTO_READINESS
            - DEPENDENCY
        insights:
          type: array
          items:
            $ref: "#/components/schemas/MetadataAnalysisInsight"
    MetadataDependencyGraphNode:
      type: object
      required: [id, objectRef, objectType, status, evidenceRefs]
      properties:
        id:
          type: string
        objectRef:
          type: string
        objectType:
          type: string
        status:
          type: string
          enum: [CONFIRMED, REVIEW_REQUIRED]
        evidenceRefs:
          type: array
          items:
            type: string
    MetadataDependencyGraphEdge:
      type: object
      required: [from, to, relationshipType, status, evidenceRefs]
      properties:
        from:
          type: string
        to:
          type: string
        relationshipType:
          type: string
        status:
          type: string
          enum: [CONFIRMED, REVIEW_REQUIRED]
        evidenceRefs:
          type: array
          items:
            type: string
    MetadataDependencyGraph:
      type: object
      required: [nodes, edges, unresolved]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/MetadataDependencyGraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/MetadataDependencyGraphEdge"
        unresolved:
          type: array
          items:
            type: object
            additionalProperties: true
    MetadataDtoReadiness:
      type: object
      required: [objectRef, status, fieldCount, reviewReasons, evidenceRefs]
      properties:
        objectRef:
          type: string
        status:
          type: string
          enum: [READY, PARTIAL, REVIEW_REQUIRED]
        fieldCount:
          type: integer
          minimum: 0
        reviewReasons:
          type: array
          items:
            type: string
        evidenceRefs:
          type: array
          items:
            type: string
    MetadataAnalysisReviewMarker:
      type: object
      required: [code, message, status, evidenceRefs]
      properties:
        code:
          type: string
        message:
          type: string
        status:
          type: string
          enum: [REVIEW_REQUIRED]
        evidenceRefs:
          type: array
          items:
            type: string
    MetadataAnalysisResponse:
      type: object
      required:
        - dbProfileId
        - mode
        - objectTypes
        - sourceProfile
        - sourceDatabase
        - targets
        - summary
        - objectInsights
        - objectProfiles
        - insightGroups
        - dependencyGraph
        - dtoReadiness
        - aiToolEvidence
        - deterministicFacts
        - reviewMarkers
        - assumptions
        - caveats
        - reviewRequired
        - blockers
        - componentInvocations
      properties:
        dbProfileId:
          type: string
        mode:
          type: string
          enum: [QUERY, TARGET]
        query:
          type: string
        target:
          $ref: "#/components/schemas/MetadataObjectIdentity"
        objectTypes:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchObjectType"
        sourceProfile:
          type: string
        sourceDatabase:
          type: string
        snapshotId:
          type: string
        collectedAt:
          type: string
          format: date-time
        targets:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchResult"
        summary:
          type: string
        objectInsights:
          type: array
          items:
            $ref: "#/components/schemas/MetadataAnalysisInsight"
        objectProfiles:
          type: array
          items:
            $ref: "#/components/schemas/MetadataObjectProfile"
        insightGroups:
          type: array
          items:
            $ref: "#/components/schemas/MetadataInsightGroup"
        dependencyGraph:
          $ref: "#/components/schemas/MetadataDependencyGraph"
        dtoReadiness:
          type: array
          items:
            $ref: "#/components/schemas/MetadataDtoReadiness"
        aiToolEvidence:
          type: object
          additionalProperties: true
        deterministicFacts:
          type: array
          items:
            type: object
            additionalProperties: true
        reviewMarkers:
          type: array
          items:
            $ref: "#/components/schemas/MetadataAnalysisReviewMarker"
        assumptions:
          type: array
          items:
            type: string
        caveats:
          type: array
          items:
            type: string
        reviewRequired:
          type: boolean
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/MetadataSearchBlocker"
        modelInvocation:
          oneOf:
            - $ref: "#/components/schemas/ModelInvocationSummary"
            - type: "null"
        componentInvocations:
          type: array
          items:
            type: object
            additionalProperties: true
...
    ErrorResponse:
      type: object
      required: [detail, code]
      properties:
        detail:
          type: string
        code:
          type: string
          examples:
            - DEPENDENCY_BLOCKED
            - P21_LIVE_PORTAL_REQUIRED_ENV_MISSING
            - P21_LIVE_PLF_UNAVAILABLE
            - P21_LIVE_PPM_REQUIRED
            - PPM_DB_ACCESS_DENIED
            - PPM_DB_NOT_FOUND